More delete fixes. Now the guadec document passes selectall- delete and undo's
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@11968 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/fmt/xp/fp_Run.cpp
@@ -4087,7 +4087,7 @@ fp_FieldFootnoteAnchorRun::fp_FieldFootnoteAnchorRun(fl_BlockLayout* pBL, GR_Gra
 	bool bRes = pp->getAttribute("footnote-id", footid);
 
 	UT_ASSERT(bRes);
-	m_iPID = atoi(footid);
+	m_iPID = atol(footid);
 }
 
 // Appears in the FootnoteContainer, one per footnote.
